src/lamplib/requirements.txt
wheel==0.40.0
pymongo==4.3.3
click==8.1.2
click-option-group==0.5.5
colorama==0.4.3
structlog==20.1.0
omegaconf==2.1.1
pyyaml<=5.3.1
requests==2.31.0
yamllint==1.15.0
shrub.py==3.0.7
black==22.10.0
setuptools==65.7.0
pytest==7.2.0
progressbar==2.5
natsort==8.4.0

numpy==1.23.5

numexpr==2.8.4

jinja2==3.1.1